+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,12 +1,12 @@
 package roomescape.controller;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
 import roomescape.dto.ReservationRequest;
 import roomescape.entity.Reservation;
 
@@ -15,21 +15,11 @@ import java.util.List;
 import java.util.Objects;
 import java.util.concurrent.atomic.AtomicLong;
 
-@Controller
+@RestController
 public class ReservationController {
 
     private final List<Reservation> reservations = new ArrayList<>();
     private final AtomicLong index = new AtomicLong(1);
-
-    @GetMapping("/admin")
-    public String admin() {
-        return "/admin/index";
-    }
-
-    @GetMapping("/admin/reservation")
-    public String reservation() {
-        return "/admin/reservation-legacy";
-    }
 
     @GetMapping("/reservations")
     public ResponseEntity<List<Reservation>> findAllReservations() {
